Moving billing business logic from consumer to the BillingLogic
